requirements.txt
telethon==1.36.0
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.4
chromadb==0.4.22
langchain==0.1.6
langchain-community==0.0.19
aiosqlite==0.19.0
python-dateutil==2.8.2
colorama==0.4.6
rich==13.7.0
schedule==1.2.1
